+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.0.17
+
  * supporting instance level inspection for states (including permissible state, see issue #54)
  * added autocreation of constants for each state ([@jherdman](https://github.com/jherdman))